+++ include/nfx/graphics/gl/pipeline/queue/RenderQueue.h
@@ -46,7 +46,7 @@ namespace nfx::graphics::gl
             }
             if (cmd.sortKey.isZero())
             {
-                cmd.sortKey = SortKey::packOpaque(SortKey::OpaqueLayer, 0, cmd.material.id, 0);
+                cmd.sortKey = SortKey::packOpaque(SortKey::OpaqueLayer, 0, static_cast<std::uint32_t>(cmd.material.id), 0);
             }
             m_commands.push_back(cmd);
         }
